Add more schemas, use black formatter

+++ lichess/schemas/GameEventPlayer.py
@@ -7,13 +7,15 @@ class GameEventPlayer(JsonDeserializable):
 
     See https://github.com/lichess-org/api/blob/master/doc/specs/schemas/GameEventPlayer.yaml
     """
+
     @classmethod
     def de_json(cls, json_string):
-        if json_string is None: return None
+        if json_string is None:
+            return None
         obj = cls.check_json(json_string, dict_copy=False)
         return cls(**obj)
 
-    def __init__(self, aiLevel: float, id: str, name: str, title: str | None, rating: float, provisional: bool):
+    def __init__(self, aiLevel: int, id: str, name: str, title: str | None, rating: int, provisional: bool):
         self.aiLevel = aiLevel
         self.id = id
         self.name = name
